refactor(core): move MessageRetriever to dedicated module
Extract MessageRetriever trait and InputMessage struct from traits.rs
to their own message_retriever.rs module for better organization.

- Create crates/core/src/message_retriever.rs
- Re-export from traits.rs for backwards compatibility
- Update memory.rs imports
