+++ crates/core/src/message_retriever.rs
@@ -1,0 +1,99 @@
+// MessageRetriever - Retrieval-only trait for conversation messages
+//
+// Design decision: MessageRetriever is retrieval-only.
+// Messages are stored via EventEmitter (message.user, message.agent events).
+// This trait provides read access for building LLM context.
+
+use async_trait::async_trait;
+use uuid::Uuid;
+
+use crate::error::Result;
+use crate::message::{ContentPart, Controls, Message, MessageRole};
+
+// ============================================================================
+// InputMessage - Input structure for message creation
+// ============================================================================
+
+/// Input message for creating a new message
+///
+/// This is the input structure for adding messages, without the ID and timestamp
+/// which are generated by the storage layer.
+///
+/// Note: Message creation happens via EventService/EventEmitter, not MessageRetriever.
+/// This struct is used by the API layer and in-memory stores for tests.
+#[derive(Debug, Clone)]
+pub struct InputMessage {
+    /// Message role (user, assistant, tool_result, system)
+    pub role: MessageRole,
+    /// Message content as array of content parts
+    pub content: Vec<ContentPart>,
+    /// Runtime controls (model, reasoning, etc.)
+    pub controls: Option<Controls>,
+    /// Message-level metadata
+    pub metadata: Option<std::collections::HashMap<String, serde_json::Value>>,
+    /// Tags for filtering/categorization
+    pub tags: Vec<String>,
+}
+
+impl InputMessage {
+    /// Create a new user input message with text content
+    pub fn user(content: impl Into<String>) -> Self {
+        Self {
+            role: MessageRole::User,
+            content: vec![ContentPart::text(content)],
+            controls: None,
+            metadata: None,
+            tags: vec![],
+        }
+    }
+
+    /// Create from a Message (useful for storing existing messages)
+    pub fn from_message(msg: &Message) -> Self {
+        Self {
+            role: msg.role.clone(),
+            content: msg.content.clone(),
+            controls: msg.controls.clone(),
+            metadata: msg.metadata.clone(),
+            tags: vec![],
+        }
+    }
+}
+
+// ============================================================================
+// MessageRetriever trait
+// ============================================================================
+
+/// Trait for retrieving conversation messages
+///
+/// This trait provides read-only access to conversation history for building
+/// LLM context. Message storage is handled separately via EventEmitter
+/// (messages are stored as events: message.user, message.agent).
+///
+/// Implementations can:
+/// - Load messages from a database (reconstructing from events)
+/// - Keep messages in memory for testing
+/// - Load messages via gRPC from control-plane
+#[async_trait]
+pub trait MessageRetriever: Send + Sync {
+    /// Get a specific message by ID
+    async fn get(&self, session_id: Uuid, message_id: Uuid) -> Result<Option<Message>>;
+
+    /// Load all messages for a session
+    async fn load(&self, session_id: Uuid) -> Result<Vec<Message>>;
+
+    /// Load messages with pagination
+    async fn load_page(
+        &self,
+        session_id: Uuid,
+        offset: usize,
+        limit: usize,
+    ) -> Result<Vec<Message>> {
+        let all = self.load(session_id).await?;
+        Ok(all.into_iter().skip(offset).take(limit).collect())
+    }
+
+    /// Count messages in a session
+    async fn count(&self, session_id: Uuid) -> Result<usize> {
+        Ok(self.load(session_id).await?.len())
+    }
+}
